[depthMap] Add new class SgmDepthList
